fix: ignore case when inferring file extension from codeSample lang prop
